# app/docs/openapi.json
{"openapi":"3.0.2","info":{"title":"SimpleAdvert","description":"Сервис для удобного размещения объявлений","version":"0.1.0"},"paths":{"/adverts/":{"get":{"tags":["Advert"],"summary":"Смотреть все объявления","description":"Смотреть все объявления.\n\n- **title**: название\n- **description**: описание\n- **kind**: вид объявления (Покупка|Продажа|Услуга)\n- **price**: цена\n- **id**: уникальный идентификатор объявления\n- **user_id**: внешний ключ пользователя, разместившего объявление","operationId":"get_all_adverts_adverts__get","responses":{"200":{"description":"Список всех объявлений","content":{"application/json":{"schema":{"title":"Response Get All Adverts Adverts  Get","type":"array","items":{"$ref":"#/components/schemas/AdvertDB"}}}}}}},"post":{"tags":["Advert"],"summary":"Разместить объявление","description":"Разместить объявление.\n\n- **title**: название\n- **description**: описание\n- **kind**: вид объявления (Покупка|Продажа|Услуга)\n- **price**: цена","operationId":"create_new_advert_adverts__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdvertCreate"}}},"required":true},"responses":{"201":{"description":"Информация о созданном объявлении","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdvertDB"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/adverts/{advert_id}":{"get":{"tags":["Advert"],"summary":"Смотреть объявление по id","description":"Смотреть одно объявление.\n\n- **title**: название\n- **description**: описание\n- **kind**: вид объявления (Покупка|Продажа|Услуга)\n- **price**: цена\n- **id**: уникальный идентификатор объявления\n- **user_id**: внешний ключ пользователя, разместившего объявление","operationId":"get_advert_adverts__advert_id__get","parameters":[{"required":true,"schema":{"title":"Advert Id","type":"integer"},"name":"advert_id","in":"path"}],"responses":{"200":{"description":"Данные объявления","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdvertDB"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Advert"],"summary":"Удалить объявление по id","description":"Удалить объявление.\n\n- **title**: название\n- **description**: описание\n- **kind**: вид объявления (Покупка|Продажа|Услуга)\n- **price**: цена\n- **id**: уникальный идентификатор объявления\n- **user_id**: внешний ключ пользователя, разместившего объявление","operationId":"remove_advert_adverts__advert_id__delete","parameters":[{"required":true,"schema":{"title":"Advert Id","type":"integer"},"name":"advert_id","in":"path"}],"responses":{"200":{"description":"Данные удаленного объявления","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdvertDB"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"patch":{"tags":["Advert"],"summary":"Обновить объявление по id","description":"Обновить объявление.\n\n- **title**: название\n- **description**: описание\n- **kind**: вид объявления (Покупка|Продажа|Услуга)\n- **price**: цена","operationId":"partially_update_advert_adverts__advert_id__patch","parameters":[{"required":true,"schema":{"title":"Advert Id","type":"integer"},"name":"advert_id","in":"path"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdvertUpdate"}}},"required":true},"responses":{"200":{"description":"Данные обновленного объявления","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdvertDB"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/adverts/{advert_id}/feedbacks":{"get":{"tags":["Advert"],"summary":"Получить все отзывы на объявление","description":"Получить все отзывы на объявление.\n\n- **text**: текст отзыва\n- **advert_id**: id внешний ключ объявления\n- **id**: уникальный идентификатор отзыва\n- **user_id**: внешний ключ пользователя, оставившего отзыв","operationId":"get_feedbacks_for_advert_adverts__advert_id__feedbacks_get","parameters":[{"required":true,"schema":{"title":"Advert Id","type":"integer"},"name":"advert_id","in":"path"}],"responses":{"200":{"description":"Список отзывов на объявление","content":{"application/json":{"schema":{"title":"Response Get Feedbacks For Advert Adverts  Advert Id  Feedbacks Get","type":"array","items":{"$ref":"#/components/schemas/FeedbackDB"}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/adverts/{advert_id}/complaints":{"get":{"tags":["Advert"],"summary":"Получить все жалобы на объявление","description":"Получить все жалобы на объявление.\n\n- **text**: текст жалобы\n- **advert_id**: id внешний ключ объявления\n- **id**: уникальный идентификатор жалобы\n- **user_id**: внешний ключ пользователя, оставившего жалобу","operationId":"get_complaints_for_advert_adverts__advert_id__complaints_get","parameters":[{"required":true,"schema":{"title":"Advert Id","type":"integer"},"name":"advert_id","in":"path"}],"responses":{"200":{"description":"Список жалоб на объявление","content":{"application/json":{"schema":{"title":"Response Get Complaints For Advert Adverts  Advert Id  Complaints Get","type":"array","items":{"$ref":"#/components/schemas/FeedbackDB"}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/complaints/":{"get":{"tags":["Complaint"],"summary":"Смотреть все жалобы","description":"Смотреть все жалобы.\n\n- **text**: текст жалобы\n- **advert_id**: id внешний ключ объявления\n- **id**: уникальный идентификатор жалобы\n- **user_id**: внешний ключ пользователя, оставившего жалобу","operationId":"get_all_complaints_complaints__get","responses":{"200":{"description":"Список всех жалоб","content":{"application/json":{"schema":{"title":"Response Get All Complaints Complaints  Get","type":"array","items":{"$ref":"#/components/schemas/ComplaintDB"}}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"post":{"tags":["Complaint"],"summary":"Оставить жалобу","description":"Оставить жалобу.\n\n- **text**: текст жалобы\n- **advert_id**: id внешний ключ объявления","operationId":"create_new_complaint_complaints__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ComplaintCreate"}}},"required":true},"responses":{"201":{"description":"Информация об оставленной жалобе","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ComplaintDB"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/complaints/{complaint_id}":{"get":{"tags":["Complaint"],"summary":"Смотреть жалобу по id","description":"Смотреть одну жалобу.\n\n- **text**: текст жалобы\n- **advert_id**: id внешний ключ объявления\n- **id**: уникальный идентификатор жалобы\n- **user_id**: внешний ключ пользователя, оставившего жалобу","operationId":"get_complaint_complaints__complaint_id__get","parameters":[{"required":true,"schema":{"title":"Complaint Id","type":"integer"},"name":"complaint_id","in":"path"}],"responses":{"200":{"description":"Данные жалобы","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ComplaintDB"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"delete":{"tags":["Complaint"],"summary":"Удалить жалобу по id","description":"Удалить жалобу.\n\n- **text**: текст жалобы\n- **advert_id**: id внешний ключ объявления\n- **id**: уникальный идентификатор жалобы\n- **user_id**: внешний ключ пользователя, оставившего жалобу","operationId":"remove_complaint_complaints__complaint_id__delete","parameters":[{"required":true,"schema":{"title":"Complaint Id","type":"integer"},"name":"complaint_id","in":"path"}],"responses":{"200":{"description":"Данные удаленной жалобы","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ComplaintDB"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"patch":{"tags":["Complaint"],"summary":"Обновить жалобу по id","description":"Обновить жалобу.\n\n- **text**: текст жалобы\n- **advert_id**: id внешний ключ объявления","operationId":"partially_update_complaint_complaints__complaint_id__patch","parameters":[{"required":true,"schema":{"title":"Complaint Id","type":"integer"},"name":"complaint_id","in":"path"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ComplaintUpdate"}}},"required":true},"responses":{"200":{"description":"Данные обновленной жалобы","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ComplaintDB"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/feedbacks/":{"get":{"tags":["Feedback"],"summary":"Смотреть все отзывы","description":"Смотреть все отзывы.\n\n- **text**: текст отзыва\n- **advert_id**: id внешний ключ объявления\n- **id**: уникальный идентификатор отзыва\n- **user_id**: внешний ключ пользователя, оставившего отзыв","operationId":"get_all_feedbacks_feedbacks__get","responses":{"200":{"description":"Список всех отзывов","content":{"application/json":{"schema":{"title":"Response Get All Feedbacks Feedbacks  Get","type":"array","items":{"$ref":"#/components/schemas/FeedbackDB"}}}}}}},"post":{"tags":["Feedback"],"summary":"Оставить отзыв","description":"Оставить отзыв.\n\n- **text**: текст отзыва\n- **advert_id**: id внешний ключ объявления","operationId":"create_new_feedback_feedbacks__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FeedbackCreate"}}},"required":true},"responses":{"201":{"description":"Информация об оставленном отзыве","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FeedbackDB"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/feedbacks/{feedback_id}":{"get":{"tags":["Feedback"],"summary":"Смотреть отзыв по id","description":"Смотреть один отзыв.\n\n- **text**: текст отзыва\n- **advert_id**: id внешний ключ объявления\n- **id**: уникальный идентификатор отзыва\n- **user_id**: внешний ключ пользователя, оставившего отзыв","operationId":"get_advert_feedbacks__feedback_id__get","parameters":[{"required":true,"schema":{"title":"Feedback Id","type":"integer"},"name":"feedback_id","in":"path"}],"responses":{"200":{"description":"Данные отзыва","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FeedbackDB"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Feedback"],"summary":"Удалить отзыв по id","description":"Удалить отзыв.\n\n- **text**: текст отзыва\n- **advert_id**: id внешний ключ объявления\n- **id**: уникальный идентификатор отзыва\n- **user_id**: внешний ключ пользователя, оставившего отзыв","operationId":"remove_advert_feedbacks__feedback_id__delete","parameters":[{"required":true,"schema":{"title":"Feedback Id","type":"integer"},"name":"feedback_id","in":"path"}],"responses":{"200":{"description":"Данные удаленного отзыва","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FeedbackDB"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"patch":{"tags":["Feedback"],"summary":"Обновить отзыв по id","description":"Обновить отзыв.\n\n- **text**: текст отзыва\n- **advert_id**: id внешний ключ объявления","operationId":"partially_update_advert_feedbacks__feedback_id__patch","parameters":[{"required":true,"schema":{"title":"Feedback Id","type":"integer"},"name":"feedback_id","in":"path"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FeedbackUpdate"}}},"required":true},"responses":{"200":{"description":"Данные обновленного отзыва","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FeedbackDB"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/auth/jwt/login":{"post":{"tags":["auth"],"summary":"Auth:Jwt.Login","operationId":"auth_jwt_login_auth_jwt_login_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_auth_jwt_login_auth_jwt_login_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BearerResponse"},"example":{"access_token":"eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJ1c2VyX2lkIjoiOTIyMWZmYzktNjQwZi00MzcyLTg2ZDMtY2U2NDJjYmE1NjAzIiwiYXVkIjoiZmFzdGFwaS11c2VyczphdXRoIiwiZXhwIjoxNTcxNTA0MTkzfQ.M10bjOe45I5Ncu_uXvOmVV8QxnL-nZfcH96U90JaocI","token_type":"bearer"}}}},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorModel"},"examples":{"LOGIN_BAD_CREDENTIALS":{"summary":"Bad credentials or the user is inactive.","value":{"detail":"LOGIN_BAD_CREDENTIALS"}},"LOGIN_USER_NOT_VERIFIED":{"summary":"The user is not verified.","value":{"detail":"LOGIN_USER_NOT_VERIFIED"}}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/jwt/logout":{"post":{"tags":["auth"],"summary":"Auth:Jwt.Logout","operationId":"auth_jwt_logout_auth_jwt_logout_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"401":{"description":"Missing token or inactive user."}},"security":[{"OAuth2PasswordBearer":[]}]}},"/auth/register":{"post":{"tags":["auth"],"summary":"Register:Register","operationId":"register_register_auth_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRead"}}}},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorModel"},"examples":{"REGISTER_USER_ALREADY_EXISTS":{"summary":"A user with this email already exists.","value":{"detail":"REGISTER_USER_ALREADY_EXISTS"}},"REGISTER_INVALID_PASSWORD":{"summary":"Password validation failed.","value":{"detail":{"code":"REGISTER_INVALID_PASSWORD","reason":"Password should beat least 3 characters"}}}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/me":{"get":{"tags":["users"],"summary":"Users:Current User","operationId":"users_current_user_users_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRead"}}}},"401":{"description":"Missing token or inactive user."}},"security":[{"OAuth2PasswordBearer":[]}]},"patch":{"tags":["users"],"summary":"Users:Patch Current User","operationId":"users_patch_current_user_users_me_patch","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserUpdate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRead"}}}},"401":{"description":"Missing token or inactive user."},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorModel"},"examples":{"UPDATE_USER_EMAIL_ALREADY_EXISTS":{"summary":"A user with this email already exists.","value":{"detail":"UPDATE_USER_EMAIL_ALREADY_EXISTS"}},"UPDATE_USER_INVALID_PASSWORD":{"summary":"Password validation failed.","value":{"detail":{"code":"UPDATE_USER_INVALID_PASSWORD","reason":"Password should beat least 3 characters"}}}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/users/{id}":{"get":{"tags":["users"],"summary":"Users:User","operationId":"users_user_users__id__get","parameters":[{"required":true,"schema":{"title":"Id"},"name":"id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRead"}}}},"401":{"description":"Missing token or inactive user."},"403":{"description":"Not a superuser."},"404":{"description":"The user does not exist."},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"delete":{"tags":["users"],"summary":"Delete User","description":"Не используйте удаление, деактивируйте пользователей.\nБан/разбан.","operationId":"delete_user_users__id__delete","parameters":[{"required":true,"schema":{"title":"Id","type":"string"},"name":"id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"deprecated":true},"patch":{"tags":["users"],"summary":"Users:Patch User","operationId":"users_patch_user_users__id__patch","parameters":[{"required":true,"schema":{"title":"Id"},"name":"id","in":"path"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserUpdate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRead"}}}},"401":{"description":"Missing token or inactive user."},"403":{"description":"Not a superuser."},"404":{"description":"The user does not exist."},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorModel"},"examples":{"UPDATE_USER_EMAIL_ALREADY_EXISTS":{"summary":"A user with this email already exists.","value":{"detail":"UPDATE_USER_EMAIL_ALREADY_EXISTS"}},"UPDATE_USER_INVALID_PASSWORD":{"summary":"Password validation failed.","value":{"detail":{"code":"UPDATE_USER_INVALID_PASSWORD","reason":"Password should beat least 3 characters"}}}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}}},"components":{"schemas":{"AdvertCreate":{"title":"AdvertCreate","required":["title","description","kind","price"],"type":"object","properties":{"title":{"title":"Title","maxLength":100,"type":"string"},"description":{"title":"Description","type":"string"},"kind":{"$ref":"#/components/schemas/Kind"},"price":{"title":"Price","exclusiveMinimum":0.0,"type":"integer"}},"additionalProperties":false,"description":"Схема для создания объявления."},"AdvertDB":{"title":"AdvertDB","required":["id"],"type":"object","properties":{"title":{"title":"Title","maxLength":100,"type":"string"},"description":{"title":"Description","type":"string"},"kind":{"$ref":"#/components/schemas/Kind"},"price":{"title":"Price","exclusiveMinimum":0.0,"type":"integer"},"id":{"title":"Id","type":"integer"},"user_id":{"title":"User Id","type":"integer"}},"additionalProperties":false,"description":"Схема для получения объявления."},"AdvertUpdate":{"title":"AdvertUpdate","type":"object","properties":{"title":{"title":"Title","maxLength":100,"type":"string"},"description":{"title":"Description","type":"string"},"kind":{"$ref":"#/components/schemas/Kind"},"price":{"title":"Price","exclusiveMinimum":0.0,"type":"integer"}},"additionalProperties":false,"description":"Схема для обновления объявления."},"BearerResponse":{"title":"BearerResponse","required":["access_token","token_type"],"type":"object","properties":{"access_token":{"title":"Access Token","type":"string"},"token_type":{"title":"Token Type","type":"string"}}},"Body_auth_jwt_login_auth_jwt_login_post":{"title":"Body_auth_jwt_login_auth_jwt_login_post","required":["username","password"],"type":"object","properties":{"grant_type":{"title":"Grant Type","pattern":"password","type":"string"},"username":{"title":"Username","type":"string"},"password":{"title":"Password","type":"string"},"scope":{"title":"Scope","type":"string","default":""},"client_id":{"title":"Client Id","type":"string"},"client_secret":{"title":"Client Secret","type":"string"}}},"ComplaintCreate":{"title":"ComplaintCreate","required":["text","advert_id"],"type":"object","properties":{"text":{"title":"Text","type":"string"},"advert_id":{"title":"Advert Id","type":"integer"}},"additionalProperties":false,"description":"Схема для создания жалобы."},"ComplaintDB":{"title":"ComplaintDB","required":["id"],"type":"object","properties":{"text":{"title":"Text","type":"string"},"id":{"title":"Id","type":"integer"},"user_id":{"title":"User Id","type":"integer"},"advert_id":{"title":"Advert Id","type":"integer"}},"additionalProperties":false,"description":"Схема для получения жалобы."},"ComplaintUpdate":{"title":"ComplaintUpdate","type":"object","properties":{"text":{"title":"Text","type":"string"}},"additionalProperties":false,"description":"Схема для обновления жалобы."},"ErrorModel":{"title":"ErrorModel","required":["detail"],"type":"object","properties":{"detail":{"title":"Detail","anyOf":[{"type":"string"},{"type":"object","additionalProperties":{"type":"string"}}]}}},"FeedbackCreate":{"title":"FeedbackCreate","required":["text","advert_id"],"type":"object","properties":{"text":{"title":"Text","type":"string"},"advert_id":{"title":"Advert Id","type":"integer"}},"additionalProperties":false,"description":"Схема для создания отзыва."},"FeedbackDB":{"title":"FeedbackDB","required":["id"],"type":"object","properties":{"text":{"title":"Text","type":"string"},"id":{"title":"Id","type":"integer"},"user_id":{"title":"User Id","type":"integer"},"advert_id":{"title":"Advert Id","type":"integer"}},"additionalProperties":false,"description":"Схема для получения отзыва."},"FeedbackUpdate":{"title":"FeedbackUpdate","type":"object","properties":{"text":{"title":"Text","type":"string"}},"additionalProperties":false,"description":"Схема для обновления отзыва."},"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"Kind":{"title":"Kind","enum":["Покупка","Продажа","Услуга"],"type":"string","description":"Вид объявления."},"UserCreate":{"title":"UserCreate","required":["email","password"],"type":"object","properties":{"email":{"title":"Email","type":"string","format":"email"},"password":{"title":"Password","type":"string"},"is_active":{"title":"Is Active","type":"boolean","default":true},"is_superuser":{"title":"Is Superuser","type":"boolean","default":false},"is_verified":{"title":"Is Verified","type":"boolean","default":false}}},"UserRead":{"title":"UserRead","required":["email"],"type":"object","properties":{"id":{"title":"Id"},"email":{"title":"Email","type":"string","format":"email"},"is_active":{"title":"Is Active","type":"boolean","default":true},"is_superuser":{"title":"Is Superuser","type":"boolean","default":false},"is_verified":{"title":"Is Verified","type":"boolean","default":false}},"description":"Base User model."},"UserUpdate":{"title":"UserUpdate","type":"object","properties":{"password":{"title":"Password","type":"string"},"email":{"title":"Email","type":"string","format":"email"},"is_active":{"title":"Is Active","type":"boolean"},"is_superuser":{"title":"Is Superuser","type":"boolean"},"is_verified":{"title":"Is Verified","type":"boolean"}}},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"anyOf":[{"type":"string"},{"type":"integer"}]}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"auth/jwt/login"}}}}}}